% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{checkNewiLabReport}
\alias{checkNewiLabReport}
\title{Scan the ilab report directory and completed report directory for
DATE_DATE_filename.csv files. Return the filename of the most recent ilab file
with a daterange more recent than the completed reports. If we're up to date,
return 'none'}
\usage{
checkNewiLabReport(ilabdir, reportdir)
}
\description{
Scan the ilab report directory and completed report directory for
DATE_DATE_filename.csv files. Return the filename of the most recent ilab file
with a daterange more recent than the completed reports. If we're up to date,
return 'none'
}
