% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{new_scanlog}
\alias{new_scanlog}
\title{This function is called by the HBIC_Scan_Reporter.sh script that is run as a cron job.
ilab path and filename are the location of the newly exported ilab report to process.
It will be scanned for scheduled scans since the last scanlog, and crossref with xnat.
config is a list with fields: admin_email and allowed_usage_type}
\usage{
new_scanlog(
  ilab_path,
  ilab_filename,
  xnat_paths,
  report_path,
  figure_path,
  config
)
}
\description{
This function is called by the HBIC_Scan_Reporter.sh script that is run as a cron job.
ilab path and filename are the location of the newly exported ilab report to process.
It will be scanned for scheduled scans since the last scanlog, and crossref with xnat.
config is a list with fields: admin_email and allowed_usage_type
}
