% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{new_xnatlog}
\alias{new_xnatlog}
\title{XNAT-only reporter}
\usage{
new_xnatlog(
  xnat_paths,
  xnat_dates,
  xnat_filename_dates,
  report_path,
  figure_path,
  config
)
}
\description{
XNAT-only reporter
}
