% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{xnat_day}
\alias{xnat_day}
\title{XNAT-only daily report. Day is date object like '1986-05-03'.}
\usage{
xnat_day(
  xnat_paths,
  day,
  figure_path,
  email_list = "",
  admin_email = "jbartolotti2@kumc.edu"
)
}
\description{
XNAT-only daily report. Day is date object like '1986-05-03'.
}
